# packages/contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: AVANA API
  version: 1.0.0
  description: AVANA API contract (Sprint 1 foundations)
servers:
  - url: https://api.example.invalid
paths:
...
  /v1/organizations/{organizationId}/courses/{courseId}/flashcards:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
    get:
      operationId: listFlashcards
      summary: List all flashcards for a course and due count
      responses:
        "200":
          description: List of flashcards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardListResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Organization or course not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/flashcards/review-queue:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
    get:
      operationId: listFlashcardsForReview
      summary: List due flashcards for spaced-repetition review
      responses:
        "200":
          description: Due flashcards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardReviewQueueResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Organization or course not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/flashcards/{flashcardId}/review:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/FlashcardId"
    post:
      operationId: submitFlashcardReview
      summary: Submit a flashcard review rating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFlashcardReviewRequest"
      responses:
        "200":
          description: Review submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitFlashcardReviewResponse"
        "400":
          description: Invalid input or rating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Flashcard not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/quizzes:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
    get:
      operationId: listQuizzes
      summary: List published quizzes for a course
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizListResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Organization or course not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/quizzes/{quizId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/QuizId"
    get:
      operationId: getQuiz
      summary: Get published quiz with questions
      responses:
        "200":
          description: Quiz details and questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/quizzes/{quizId}/attempts:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/QuizId"
    post:
      operationId: submitQuizAttempt
      summary: Submit answers for a quiz attempt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitQuizAttemptRequest"
      responses:
        "200":
          description: Attempt scored and submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitQuizAttemptResponse"
        "400":
          description: Invalid input or answers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/quizzes/{quizId}/attempts/{attemptId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/QuizId"
      - $ref: "#/components/parameters/QuizAttemptId"
    get:
      operationId: getQuizAttempt
      summary: Get a specific quiz attempt result
      responses:
        "200":
          description: Quiz attempt result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Quiz attempt not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/study/analytics:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
    get:
      operationId: getStudyAnalytics
      summary: Get aggregated study analytics for course
      responses:
        "200":
          description: Course study analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyAnalyticsResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Organization or course not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/study/recommendations:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
    get:
      operationId: getStudyRecommendations
      summary: Get study recommendations for course
      responses:
        "200":
          description: Course study recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyRecommendationsResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Organization or course not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  securitySchemes: {}
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    FlashcardId:
      name: flashcardId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    QuizId:
      name: quizId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    QuizAttemptId:
      name: attemptId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: ["error", "request_id"]
      properties:
        request_id:
          type: string
        error:
          type: object
          additionalProperties: false
          required: ["code", "message"]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - unprocessable
                - internal_error
            message:
              type: string
            details:
              type: object
              additionalProperties:
                oneOf:
                  - type: string
                  - type: number
                  - type: boolean
                  - type: integer
                  - type: null
...
    FlashcardRating:
      type: string
      enum:
        - again
        - hard
        - good
        - easy
    FlashcardResource:
      type: object
      additionalProperties: false
      required:
        - id
        - organization_id
        - course_id
        - document_id
        - question
        - answer
        - card_type
        - difficulty
        - due_at
        - interval_days
        - ease_factor
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        generated_content_id:
          type: string
          format: uuid
          nullable: true
        question:
          type: string
        answer:
          type: string
        explanation:
          type: string
          nullable: true
        card_type:
          type: string
        difficulty:
          type: string
        due_at:
          type: string
          format: date-time
        interval_days:
          type: integer
        ease_factor:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FlashcardListResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "flashcards", "next_review_count"]
      properties:
        request_id:
          type: string
        flashcards:
          type: array
          items:
            $ref: "#/components/schemas/FlashcardResource"
        next_review_count:
          type: integer
    FlashcardReviewQueueResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "due_cards"]
      properties:
        request_id:
          type: string
        due_cards:
          type: array
          items:
            $ref: "#/components/schemas/FlashcardResource"
    SubmitFlashcardReviewRequest:
      type: object
      additionalProperties: false
      required: ["rating"]
      properties:
        rating:
          $ref: "#/components/schemas/FlashcardRating"
        reaction_ms:
          type: integer
          nullable: true
    SubmitFlashcardReviewResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "success"]
      properties:
        request_id:
          type: string
        success:
          type: boolean
...
    QuizResource:
      type: object
      additionalProperties: false
      required:
        - id
        - organization_id
        - course_id
        - document_id
        - title
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: ["draft", "published"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    QuizDetailResource:
      allOf:
        - $ref: "#/components/schemas/QuizResource"
        - type: object
          additionalProperties: false
          required: ["questions"]
          properties:
            questions:
              type: array
              items:
                $ref: "#/components/schemas/QuizQuestionResource"
    QuizListResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "quizzes"]
      properties:
        request_id:
          type: string
        quizzes:
          type: array
          items:
            $ref: "#/components/schemas/QuizResource"
    QuizResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "quiz"]
      properties:
        request_id:
          type: string
        quiz:
          $ref: "#/components/schemas/QuizDetailResource"
    QuizQuestionAnswerInput:
      type: object
      additionalProperties: false
      required: ["questionId"]
      properties:
        questionId:
          type: string
          format: uuid
        answer: {}
    SubmitQuizAttemptRequest:
      type: object
      additionalProperties: false
      required: ["answers"]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestionAnswerInput"
    QuizAttemptResult:
      type: object
      additionalProperties: false
      required:
        - attemptId
        - quizId
        - score
        - correct
        - total
        - answers
        - completedAt
      properties:
        attemptId:
          type: string
          format: uuid
        quizId:
          type: string
          format: uuid
        score:
          type: number
        correct:
          type: integer
        total:
          type: integer
        answers:
          type: object
          additionalProperties: true
        completedAt:
          type: string
          format: date-time
    SubmitQuizAttemptResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "attempt"]
      properties:
        request_id:
          type: string
        attempt:
          $ref: "#/components/schemas/QuizAttemptResult"
...
    QuizAttemptResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "attempt"]
      properties:
        request_id:
          type: string
        attempt:
          $ref: "#/components/schemas/QuizAttemptResource"
...
    StudyAnalyticsResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "analytics"]
      properties:
        request_id:
          type: string
        analytics:
          $ref: "#/components/schemas/StudyAnalytics"
    StudyRecommendationResource:
      type: object
      additionalProperties: false
      required: ["id", "summary", "topics", "source"]
      properties:
        id:
          type: string
        summary:
          type: string
        topics:
          type: array
          items:
            type: string
        source:
          type: string
          enum:
            - accepted_lesson
            - flashcard_review
            - quiz_attempt
            - recommendation
    StudyRecommendationsResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "recommendations"]
      properties:
        request_id:
          type: string
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/StudyRecommendationResource"